Add Perl builds using the category:
Now we read the category to tell us how to build it. We also have
a Perl build which will build a Perl module. There is a lot to
say about this and a lot of "except for" or "oh, and".

Here are some of the things to note:

* Our builds should be "defaults", while we should allow overriding
them in packages. Perhaps we should allow multiple ways
("Makefile", "CMake", etc.).
* It's now doing a stupid switch. This should be removed in favor
of a proper type check with a 'build' method that knows how to
build it. We'll get there.