Add preliminary bundling capability:
This commit adds the ability to bundle the application at the end
of the build cycle. There are a lot of gotchas and FIXMEs which
will be resolved in the upcoming commits (and will probably be
raised as tickets as well) but the main feature is technically
now available.

This commit adds information on how it works and some of the
gotchas to follow-up on.

We now have "--output-dir" (default to '.', which probably isn't
a good idea) and that allows us to, once we finish building an
application, to move everything there in order to bundle it.

Theoretically the bundling can be done using rsync(1), but we need
to still know which files in a build cycle were created by which
package we bundled. We decided to go simple and reliable and just
scan the BUILD directory at the end of a build cycle. This tells
us what files were added. (None should be changed or removed! We
cannot detect a change now, only a remove, and we die on it.)

We then copy all the files from "main" to the output directory for
bundling. A nice feature idea is adding custom filters per app of
what should be copied and what are artifacts that can be ignored.

When copying we resolve all the files and the symlinks and we
recreate the symlinks. We also do the same with the permissions,
so any executables stay that way. This is similar to what rsync(1)
does if used with "-a" (except we don't support special devices
and extended permissions - we don't care about that).

After finishing the recreation of the diff, we create a xz tarball
file and copy it to a new directory under the category and package
name in the output directory.

This isn't yet a fully working solution because of the following
items:

* We don't mark the files we found yet, so we will technically
  create an archive of everything - that needs to be fixed
* We should clean up the output directory when we're done
  creating a bundle
* We should add the version to the archive
* We should add metadata in Perl packages about the perl it
  was compiled for - perhaps even adjust the directories
  accordingly?