Introducing Pkt::Builder:
The build up until now was in the command line application. It is
now moved to a proper class for building the package. The code is
mostly the same. We removed the build prefix and use Path::Tiny's
integration with File::Temp to creat the temp directory.

All other work on th build system needs to be rebased on top of
this.