Move bundling into a Bundler class:
A bit of a clean up. We now have a proper bundler as Pkt::Bundler.
It is created and used by the Pkt::Builder object. It allows a
user to provide a different bundler (must be a Pkt::Bundler for
now) or to provide the arguments for the bundler. This allows us
to have defaults, user arguments, and overrides for those, while
still allowing the user to provide a completely different bundler.

It would now be easier to do the diff between builds.