Clean up after the build process (optionally):
Two additional items were to clean up the main build directory
(BUILD-XXXXXX) and the bundle directories (BUNDLE-package).

We also introduce "--keep-build-dir" to prevent removing the build
directory which can be useful when wanting to build against an
existing build directory as part of a new build. This is the
counterpart to the "--build-dir" option.