Use Path::Tiny for the tempfile:
Now we don't need to delete the tree ourselves.

We should also double check if we can't do the same with the
BUILD directory. If a build fails, it will keep the BUILD dir.
We can either add a DESTROY() or use Path::Tiny's tempdir shortcut
which has a guard.