Fix copy location, work around filesystem mismatch:
The bundle location was assuming on the current directory, which
was wrong. It now uses the `bundle_dir` attribute.

The build directory and the output directory might be different
filesystems (such as the build running in /tmp being a tmpfs
filesystem, while the output directory being a normal filesystem)
and then *nix gets upset and fails to move the file over.

The reason it fails is because it uses the `rename()` mechanism
to move the file, which is a filesystem node update. This won't
work between filesystems, duh.

Instead we copy it over and delete the original.