Remove newline in the log messages:
The logger itself should decide whether to add a newline or not,
depending on where it logs to. If it logs to the screen, it can
add a newline. If it writes to MySQL or Syslog, it won't add a
newline. The newline is part of the representation of the message,
not the message itself.