Cleanup "open()" a bit:
Since we're using the logical "or" operator, there is no need for
parenthesizing the "open()" statement. It will run everything to
the left of "or" and only then what's to the right of it.

Also, an entire line for everything is too long. We can just put
the "or" on the next line with one level of indentation. Much more
readable.