We're already using Path::Tiny, not Cwd:
We removed all `Cwd` usage with `Path::Tiny` instead. Notice that
`Cwd` isn't actually loaded here anywhere.

The only reason we can use it is because `Path::Tiny` loads it
into memory and we call it by the fully qualified subroutine name
(`Cwd::abs_path`). This is, in fact, a hidden bug. If we removed
`Path::Tiny`, this would crash.