Include the error code in the output:
When Perl fails to do something, it keeps the errors in two
variables: `$@` and `$!`. The first stores Perl interpreter
errors. The second stores the OS errors.

Since opening a file is an operating system operation, if we
can't open (or close) a file, the error will be stored in the
OS error variable (`$!`).