Added a ConfigReader Package which can be made to adapt with and fetch config from any config-file in any storage
ConfigReader.pm takes in a 'type' of the config-file and an array ref with holds the arguments needed to fetch config from the config-file. It instantiates an object of required Package (Adapter Class) automatically, which knows how to fetch the config, given necessary information in args. The Adapter class must have the method get_config(), and to ensure that, it uses the Role Pakket::Role::ConfigReader.

The get_config() method of the ConfigReader package returns a hashref with the fetched information of designated configuration.