Changed the way of accessing the configuration file, now using ConfigReader
Previously, all the configuration was fetched from the .TOML
file using the TOML::Parser. Now it uses an Object of ConfigReader
package to get it. ConfigReader object automatically determines
the way of getting the configuration file from the 'type' it gets,
and it uses the elements of 'args' for necessary information
needed to get the config.