Fix: Use INSTALL_BASE, not PREFIX:
There was a problem when using the `PREFIX` configuration option
with some modules, warning about not being able to use both this
and `INSTALL_BASE` which we didn't set.

To counter that, we would explicitly set `INSTALL_BASE` to empty
all the time. This was not the right decision. When we started
using `System::Command`, those warnings returned and I had no idea
how to remove them.

Instead, after consulting with Leon T., I decided to use
`INSTALL_BASE` directly and remove `PREFIX`. In Leon's words:
"`INSTALL_BASE` is `PREFIX` done right".