Add a logger role:
We now have a proper logger role that adds default logging to
everything, but also a logger object that you can set yourself.

We're using Log::Contextual and Log::Dispatch under that.

It's a bit tricky, but what's going on is that Log::Contextual
requires you to call `set_logger` to set the current package
logger. We want the role to do this but it can't, because if it
will, the logger will only apply to the Role package, not the
consumer.

Due to this we're basically setting it in the role by calling it
from the consumer. The problem is that the consumer needs to have
it, which we can't force it to import. (Technically, we could but
then we won't have the compile-time functions "log_info" and
family.)