Use proper Log::Contextual pattern:
I like and dislike Log::Contextual.

The proper with the previous pattern is that once the role is
used, it will call "set_logger". Once some other package consumes
the role, it will set the global logger again. Log::Contextual
doesn't like it (and it is silly).

What we do instead is have a subclass of the logger which has
the defaults, and the CLI then sets the logger manually once it
has enough information.

This means several things:

* Whowever wants to log loads the logger. That means they now
  have access to the global logger, including all the levels we
  want. They don't have to ask for those levels or for the logging
  subroutines. They already get those from the Log class.

* If there is no global logger already set, the Log class will
  set the global one for you.

* The first point of entry needs to set the global logger itself.
  This actually isn't that fun, because we might have several
  points of entry, such as multiple App::Cmd commands, implemented
  in multiple classes. We'll try to refactor this.