Add init command:
The init command allows to initialize a proper Pakket installation.
This creates a directory with a bash script that allows you to
seamlessly connect to a Pakket repo, installed either globally
or locally.

There is still a limitation here that you can only have one
Pakket installation - either global or local. This should be fixed
to allow both global and local. The local will be loaded first,
the global after.