Add script to generate spec files from MetaCPAN:
This script takes either a distribution or module and creates a
distribution spec file. If it's a module, it will retrieve the
distribution in which it is included.

The output is currently only in TOML (this should be an entire
importer module, probably) and takes into account the different
states:

* runtime:
    These are dependencies to install in order to run the module.
    (this includes build-time because it will require it for
     tests.)
* test:
    These are dependencies to install during testing. If the build
    process does not run tests, there is no need to install these
    modules.
* configure:
    These are dependencies to install during build