Remove versioning entirely from config scaffolder:
The versioning didn't really work right. Modules and distributions
can have different versions and you can have a module with version
X inside a distribution with version Y.

It would search for a distribution by the module name and then add
the module version on that, and search for it. That's okay if they
fit, but that's not correct.

After enough experiments, I resolved with "screw this for now".

If you have a hard-coded prereq, it should be added after
scaffolding your configuration.