Add a tool to create a package index:
We're changing the directory structure to support versioning
better. This means that we will need to know which version is the
version to be used. We need an index.

This tool scans a directory of specs (generated from the tool that
generates the packages' specs), parses all versions to determine
the latest one, and creates a proper JSON index.