Ignore "perl" as a dependency in category "perl":
This is a technical limitation that we will need to clean up later
to be used as a per-category issue. Meanwhile we're just
hard-coding that we shouldn't run "perl" in "perl".

It would appear because modules can include "perl" as their CPAN
dependency (meaning a certain version of "perl"). But for us,
"perl" is under the "system" category.