Move LD_LIBRARY_PATH to better location:
While we *were* setting it using `local $ENV{'LD_LIBRARY_PATH'}`,
it is better that we only use it in env. This makes a different
comment in the code much more readable.

As an additional improvement, we do not remove any existing
`LD_LIBRARY_PATH` entries. We just prepend to it.