Add /ms flags to regular expressions:
On /m:

'^' and '$' match the beginning and end of a string, not a line. This means
that when you have a string that contains newlines, these anchors will not
work properly. Using /m changes is so they work for lines, even within a
multi-line string.

On /s:

While '.' usually means any character except a newline, with /s, it also
matches a new-line.

From perlre:

    Used together, as "/ms", they let the "." match any character whatsoever,
    while still allowing "^" and "$" to match, respectively, just after and
    just before newlines within the string.