Clarify fancy (but confusing) App::Cmd syntax:
App::Cmd works by commands declaring themselves as the following:

    use Main::Module -command;

Perl::Critic thinks that "command" is a subroutine (even though
it isn't). This is confusing to beginners as well.

It is, in fact, a string. So now we're quoting it to clarify what
this syntax actually means.

Interestingly, even then Perl::Critic complains, because we're
mixing a minus (-) with a string ("command") and it assumes this
is a mistake (which is most cases it will be), so we're also
locally disabling that policy.