unbreak the scaffolding script that was generating wrong configs
the script was using "dependency type" (e.g. requires, recommends, etc.) in
place of "dependency phase" (e.g. configure, runtime, etc.)

so we had

    [Prereqs.perl.requires.Data-Dumper]
    version = "2.154"

instead of

    [Prereqs.perl.runtime.Data-Dumper]
    version = "2.154"

since we recognise only a limited number of phases, and "requires" and friends
are not among them, this made Pakket skip all dependencies when building
modules