Crappy hack to avoid dying when scan directory contnet that still exist:
When unpacking, we would scan the directory. If we're already
using a build directory which has had a package unpacked inside it,
the scanning fails because it doesn't detect new files created
by the unpack.

We had a hack of a flag allowing us to ask it not to die.

The best solution would be to split the scan to a version that
dies and a version that just unpacks and creates a snapshot.