Finally break apart builders to multiple classes:
This has bugged me for a while. All the builders are now proper
classes, based on category.

There is a role that allows all builders to make sure they can
run commands and generate the proper environment variables.

Perl contains the code for both ExtUtils::MakeMaker and
Module::Build, while System delegates to a Makefile based one,
so it could also allow CMake and SCons when we encounter them in
the future.

What bugs me is that we fetch configure flags (which might or
might not exist) and then send it everywhere. Instead I would
prefer to have a general hash of metadata in which the configure
flags would be one possible field.

I'm still not sure how to introduce additional categories without
loading them in the Builder.pm object and adding them to the
"builders" attribute. Two possibilities exist:

1. Make this automatically load with Module::Pluggable, but that
   adds a problem when newer versions possibly remove outdated
   builders. They might remain from older versions and `cpan(m?)`
   doesn't remove before upgrading.

2. Add a configuration option to include additional ones, which
   makes more sense.

This closes GH #12.