Reuse installation directory:
Instead of a fresh install of the modules (and dependencies), we
start from the current installed directory.

We could now keep several previous copies and even revert to an
older installation directory. That part isn't written yet.

The symlink part should be optional in the future because this
wouldn't work on Windows.