Remove old directories automatically:
We now clean up the old directories, but allow keeping as many
as you want. Default: 1.

If you define this as anything under zero, it turns into 1.

When it is set to 1, it removes anything but last, which is why
it makes a good default. Storing other amounts allows us to add
the ability to revert to them.