Add constants, rename parcel's files dir:
The file directory used to be the same as the filename which was
basically the package name plus the version minus the extension.

This made the code ugly. It's also awkward and requires parsing
in order to figure out how to do this properly.

So now we're 'files' as the directory name. Also, it's been moved
to a constants module.

We then move more stuff to constants.