Introduce the Package object:
This is still a work in progress but here's the first draft. It's
meant to introduce a part of it without breaking anything.

First we have the Pakket::Package class which provides the
basic attributes for name, category, and version.

Then we add the requirements (config, test, runtime, and the
general "prereqs" which contain everything).

We also introduce the versioning as part of the Package class, and
it determines by itself what the appropriate versioning scheme is.

This removes the 'versioning.toml' file. That approach had two
flaws: 1. It requires file-based configuration, which won't work
once we moved outside directory-based structure. 2. It means that
a package does not contain the knowledge of how to do diffs for it.

For the majority, the rest of the code still uses about the same
code, but passes the package object between methods and uses that
instead. Since the versioning scheme is now part of the Package
class, that had been cleaned up.