Redo the versioning schemes:
One thing I forgot was the versioning schemes. Perl has a sort of
SemVer. NodeJS has SemVer, but Perl's version is more common in
native packages.

Instead they're reimplemented with names "Default" and "SemVer",
so they could be reused by any category or package that wants it.

There was another level of indirection between
Pakket::Version::Requirements and the schemas
(Pakket::Version::Schema::*) which was removed by adding the methods
of Pakket::Version::Requirements to the Schema classes (now in
Pakket::Versioning::* classes).