Add "system" category explicitly:
The existing (read: one replaced by this branch) versioning scheme
code broke all system package builds by not providing "system" as
an accepted category with a known versioning scheme.

This is now implicitly fixed by this new implementation because it
has a fallback of the default (to the "Default" versioning scheme,
which is what Perl uses).

This commits also adds "system" as an explicit category.