Fake usage of prereq as package to make it work for now:
Until now we just passed strings around, which allows (or made?)
us to treat prereqs and packages as the same thing. The first time
we call 'run_build', it's called with a package. Then it's passed
with strings that match a package, but are actually prereqs.

For now we're going to fake the prereq sent to it as a package
object. Later we will have to split them apart.