Make install CLI use package object, remove exact_version:
The "install" command line class needs to use the package class,
but it had assumed that when a user was providing a version, it
means it's the requirement exact version (which led to the
previous hackery with "exact_version" in the package and extra
parameters).

This cleans it up and removes it entirely. The drawback is
basically undoing cb492384570b635ea44f0b7a804b3a1789a60c93. We
will need to return it in the future once we have an understanding
of what's a package and what's just a prereq.

General thoughts: Let's make it `pakket install cat/name@version`.