Fix versioning comparisons:
We are dealing with various forms of versions. Some of them are
v-strings, some of them are not. Some dotted decimals, some just
decimals.

Just for fun:

    using version.pm's parse()
    "2.49" -> "2.490"
    "2.49.1" -> "2.049.1"
    "v2.49" -> "2.049000"

    using version.pm's qv()
    "2.49" -> "2.049000"

Now ww're just normalizing with qv() turning everything into a
versioon object and then numify it.

This fixes versioning on modules like Getopt-Long which have a
v-string version.