Correctly locate the parcel file:
The pracel file was being searched using the location of the original
directory, which isn't necessarily the same name as our parcel.
Our parcel includes our own composition of name-version.pkt, while
some upstreams provide their code in directories that either wholly
different or similar but not identical.

Until now we were able to get away with it because we only tested
on well-behaving Perl and C libraries. No more. :)