Add Pakket::Requirement, no other change:
We now have a proper requirement object. It's probably rough around
the ages and we'll see breakage once we start using it.[1]

You can create a new requirement with:

    my $req = Pakket::Requirement->new(
        'category' => 'perl',
        'name'     => 'Plack',
        'versions' => [ @available_versions_from_index ],

        # Either exact version:
        'exact_version' => '1.200',

        # Or version range:
        'version_range' => '>= 1.200, != 1.300',
    );

Now you can get the latest version:

    my $latest = $req->latest_version();

It will retrieve it based on either an exact version provided
(which is a shortcut checking if that version exists) or on a
range (which checks it using the appropriate implementation).

The implementations only need to implement two methods now:

* sort_candidates
* get_latest_version

They must determine how they sort them and how to get the latest
version. The `get_latest_version` method needs to implement the
range implementation, if it can. The `sort_candidates` implements
the parsing in order to compare versions.

[1] Testing, anyone?