Add new Pakket::Requirement, not used yet:
This requirement class allows us to represent a requirement and
differ between it and a package.

Up until now we've basically treated prereqs as a package. We
used version 0, as if it were a package version, instead of a
requirment for any version. We treated the package version as
either a version or a minimum requirement, which caused a lot of
confusion. This cleans it up.

The Requirement object can:

* Represent a category
* Represent a name
* Represent a version.

The version is currently only an exact version. This is to move to it
incrementally. In the future, with GH #75, we will make this also
support a range of version requirement.