Correct a typo:
We had used 'prereqs' instead of 'Prereqs'. Because of this, when
the Installer tried to access the prereqs, it couldn't. This meant
it didn't find and install *any* prereqs for anything.

When building a lot of things at once, this was unnoticeable, but
once in a while we would get "Can't find ExtUtils::Config" - this
is why.