Add FatPacked version of the Seacan builder:
This allows you to run the Seacan build without building it first,
which is much easier. Until now we had three levels of
bootstrapping, so at least that's one level less.

This does mean we will have to regenerate this every time we update
the script. That might be something to put in a pre-commit hook.