Move the index to the Repository Backend class:
We had previous assumed each Repository will contain an index which
is only created by the Backend object. However, it's much simpler for
the Backend to decide how to store the index and how to fetch
information when the Repository wants.

This allows a DBI-based backend, for example, to request each time
instead of requesting everything at first and keep it in the index.