Implement improved Repository setup:
We have redesigned the Repository setup.

* We first have the custom repository objects:
  * Pakket::Repository::Config
  * Pakket::Repository::Parcel
  * Pakket::Repository::Source

Those Repository objects are in charge of knowing about specific
input and output and are not generic in any way.

* Then we have the general Repository object:
  * Pakket::Repository

This object has a backend of Pakket::Repository::Backend and it
provides several methods which it delegates to the backend,
specifically:

* all_object_ids
* retrieve_content
* store_content
* retrieve_location
* store_location

Each backend implements the above methods, allowing each custom
Repository to access them using the Repository API.

Each method works with an ID, which the Backend decides how to
use.

We implemented the Backend::File, allowing storing files and
contents in files. It hashes content (and files) in order to have
a consistent ID to value.

We also implemented the Config repository, which stores content of
configurations from packages and inflates back configuration from
the repo.