Add Parcel and Source Repository objects:
We now have the ability to store parcels and sources.

The parcel repository is a simple location-based repo, the source
is a bit more complicated.

The Source repository receives a path, assuming it's a directory
(although it could also be a file, of course). It archives it
and then stores that archive.

This allows the source to either be a tarball (as RPM likes) or
an open directory (the original design for Pakket).