Move the Bundler to use Repository::Parcel:
This is so much fun. :)

The Bundler should now be using the new Parcel repository. This
removes the need to create the parcel file ourselves. We instead
provide it with the new bundled directory. It creates a file from
it and stores it in the backend using the Repository.

There is more room for cleanups, left as comments for now.