Fix installer:
Other than an unnecessary check (which was also wrong), the installer
would recursively extract into the same temp directory of the
original parcel that was extracted at first. This caused a permission
error and other crazy stuff.

Clean up as soon as we unpack, then continue to unpack each parcel
dependency separately.