Sync install cache with all installed items:
We have two caches: A build cache and an install cache. The build
cache is meant to prevent rebuilding the same thing in recursive
cycles. The install cache tries to do the same, but for recursive
installations.

Anything that is installed will be added to the build cache
automatically, but only because before installing we try to build
it. (The installation is triggered from the building.)

Here's the problem: If the installation has to install
dependencies, because they don't go through the build process
(but instead installed via the installer alone), they only appear
in the install cache, but not in the build cache.

We then need to synchronize the build cache with the install cache
once we finished installing something.

This isn't a complete solution, it's just the first action.