Fix (re)creation of active symlink
When the active symlink already existed, we were trying to make it point
to the new directory by calling $link->move($new_dir), which actually
renames the link itself, instead of repointing it.

So now we simply delete the symlink when it exists; then we just create
the symlink pointing to the proper place.