Remove duplication in retrieving and removing locations:
Both the Parcel and Source repository retrive locations and remove
them, so they share exactly the same code for it. One thing we
could do is have their user call the upper method of retrieving
location, but I think it's not a good separation of concerns and
we won't have a more accurate logging from it.

Instead, I'm currying this into a general request that also logs
the appropriate action from it.

Another benefit is that now Parcel repository class will log a
removal properly (which we didn't notice did not exist in the copy).