Rename PrintableNames to BasicPackageAttrs:
PrintableNames was correct for what we had, but if we also introduce
a non-printable ID for a package (this will happen when we add
arbitrary attributes, which will part of the ID but not part of the
package name), then PrintableNames will no longer be correct.

I think it's correct to say the Requirement has basic Package
attributes.