Use Package/Requirement ID when necessary, not full name:
This is the second part of the distrinction between the ID and
the printable name of a package or a requirement.

When retrieving, storing, or removing locations and content, we
use the ID and not the full name. The full name is printable, while
the ID is able to include non-printable values which determine a
distinct ID for the package/requirement, such as attributes.