Use the new has_object instead of inquiring manually:
We can now use the existing has_object in order to check whether
a certain package exists or not.

We are creating a Pakket::Requirement object on the spot, but I
think in the future we could have created them ahead of time in
perl_bootstrap_modules, but we'll get there in the future.