Add Pakket::Config and HasConfig role:
This is the starting step of introducing a global configuration.

The challenge is to remain flexible and making it easy to write
tools that can use the configuration without calling the CLI
commands.

The Config class is flexible, allowing to control the directory
(default to ~), the prefix (default to .pakket), and the extensions
(default to JSON, YAML, and standard cfg). You can also set the
full list of exact config files (if you want to use user-supplied
config file paths). The Config class simply reads the configuration.

The HasConfig role provides a "config" attribute which is populated
from what the Config object has read.