Move install CLI and Pakket::Installer and Parcel Repo to config:
This is hard to follow. I think it's easier to simply view the
install.pm file to understand that one.

As far the Installer:

* We use HasConfig to get a config attribute.
* We add parcel_repo_backend attribute to hold the backend. This
  should eventually be in the HasSpecRepo role.
* We remove the parcel_dir because it's folded into the
  parcel_repo_backend attribute.

In the Parcel Repository:

We remove the requirement for a directory (since it was just
hard-coding an attribute for a File backend, not allowing us to
provide *any* backend) and no need the default to build the backend.