Convert build command and Builder and Bundler to use config:
The build command now checks each directory it receives, if it has
received them. The rest is cleanups.

One thing to take into account is that the Bundler would use an
absolute path of the output dir in order to make sure it copies to
the right one even though it chdirs outside it. This *might* be a
problem.

In order to hopefully adjust correctly, I added a BUILD method to
inflate the repo object. I don't think it's necessarily safe since
the object might internally lazily evaluate the directory.

This needs to be tested. Who knows, it might even work without it.