Cleanup attempts to retrieve latest version:
We had an ad-hoc attempt to retrieve the latest version since there
is still no comparison operators and abilities to sort versions
yet. This turned into a weird grep operation, hard-coded in the
wrong place.

Instead, we now have a nice method. This will allow us to control
the usage of the version (for the release data feature) and to
provide a comfortable place to add the comparison operators work.