Fix bug: Get version correctly:
When we retrieve the object IDs, it's not the objects, but just
the IDs. We still need to grab the version from it.

There are two reasons I don't like it. The first is that we're not
getting objects, but IDs instead. Secondly, it's strings that we
need to parse in order to get the relevant information from them.
It's just lucky for us that the ID is in the string, sort of. If
it weren't, we wouldn't be able to get the version information.

A better approach would be to ask for a specific package from
object. We don't have that, but it would be good to have it.