Fix: Implement our own "impolite" check:
The Archive::Any impolite check is not enough for us to know how
we can use the archive content, because we cannot trust the files
list. It might contain a file in a directory or a directory.

Here is an example of a file list which works for us:

    [
      'Safe-Isa-1.000006/',
      'Safe-Isa-1.000006/Changes',
      'Safe-Isa-1.000006/lib/',
      ...
    ]

Here is an example of a file list that confuses us:

    [
      'HTTP-Headers-Fast-0.21/Build.PL',
      'HTTP-Headers-Fast-0.21/Changes',
      ...
    ]

The latter example will confuse us as we would be serving the first
item thinking it's the directory of the files, when it is actually
a full file path. This is probably due to different file types.

The solution is to create our own which runs after extracting,
inspecting the directory to decide whether the archive as polite
or not, and whether it included a single directory or not.