Cleanup: Only print 'export's once:
This is annoying that we're printing the environment variables
every time, so I changed it to only print once. The environment
variables aren't changing, so why print it every time? Too much
output.

I also added a small index of the different logging levels, just
because I'm sick of looking it up every time.