Add log_success and log_fail helper logging functions:
I couldn't find any better way to do this, because Log::Any and
Log::Dispatch do not provide enough control over it. If they allowed
sending more than just a message, or they would strip ANSI color
codes, it would have been much easier. Also Term::GentooFunctions is
not flexible enough. Ugh. Fine.

What we do is basically provide two additional functions which will
log to the screen either using Term::GentooFunctions or send the
equivalent to any non-Screen logger.

This can fail in a bunch of ways that do not affect us. A cleaner
way might be to use a Log::Any::Proxy but... ENOTENOUGHTIME for that.