Fix: Handle multiple prereqs:
When using the --input-file option, we provide multiple packages.
Those call ->build every time. The build has side-effects of
setting up bootstrapping, so bootstrapping happens for each package
we want to build. Damn.

This calls ->build only once, and then calls run_build, which will
not call bootstrapping.

This should be done in a different way, but meh.