Read config before setting logger:
We want to allow the config to specify the logger, so we need to
read it before.

The config reader has only two logging actions: One info() to say
what it's reading (we can ignore that) and one critical() in order
to croak when failing. The croak() will receive the result of the
critical() call so it will still croak() with the appropriate
message, so it's fine.

(This *does* mean that when the config reader dies during the
build command - the beginning of it! - it will not log that to any
file.)