Creating PackageQuery:
During discussion, we realized that we need to represent three
types of package-related objects, but only use two classes.

We currently have Pakket::Package and Pakket::Requirement. The
Package class represents an existing package with its configuration,
prereqs, and so on. Requirement class represents a specific package
we would like to use as a requirement, but without all the package
information.

What we need instead is:

* An object representing the package, with all the details.
  (Pakket::Package)
* An object representing the requirement, but not in a specific
  version. Instead, it uses a range. (This *should* be
  Pakket::Requirement.)
* An object in between, that represents what we need as a package
  but without the details. This is what Pakket::Requirement
  currently gives us. We need this in order to do queries, so the
  name Pakket::PackageQuery was picked.

This introduces this class and uses it here and then. I will then
move to clean up the Requirement class and usages to contain
ranges and remove the need to create a Requirement when we in fact
want a PackageQuery object.