Scaffolder: fix is_local to work with deps
This change will allow to use it separately from --no-deps
and will add the read requirements found in the local package's
META.json/META.yml files.

The --is-local flag now takes mulitiple values of all
package names you want to include from local directory
(so that they can be added as deps of each other).

Note: finding the version of a local dependency still requires
some work (next task).
for now we just list the matching files with the same
prefix and pull the last one (meaning you better not have
multiple versions of a local module in the cache dir).