Use Makefile.PL as failover instead of Module::Build:
We found cases of Perl modules that have both Build.PL and
Makefile.PL, which means that (unless they're in the exception
list) they should be using Module::Build, but they do not have
Module::Build defined as a depenedency, so it fails.

We now also check for a Makefile.PL and try to fallback on that.