Move dot to the end of @INC:
One of our problems with building Perl packages is that dot
(current directory) is required for many package builds but you
cannot know for certain it exists. (Perl 5.26.0 removed it from
@INC and toolchain now adds it back.)

When we use PERL5LIB to add it again, it adds it in the beginning.
This breaks any badly-formed distributions which include their lib
pm files in the main directory instead of in the lib/ directory.
They are accidentally loaded because of the dot in the beginning.

In order to fix it, we're adding @INC to the beginning of PERL5LIB,
effectively moving the dot to the end of it.