Change order, localize PERL5LIB:
PERL5LIB needed to be localized to prevent the environment of the
builder from being injected into the internal process that fetches
the @INC. This is possibly unnecessary with Role::RunCommand.

We changed the order so we first load main/lib in case pre-built
packages are available, then @INC, then dot.