Don't die if do not find distribution_name for module_name
Just temporary use name with dashes instead of colons s/::/-/.
It will simplify things with dependencies for local(non-CPAN) modules.

In case if dependency is correct and exists in spec repo, everything
will work out of the box without creating Pakket.json with
module->distribution map.

In case if dependency doesn't exist, we will die on the next step when
trying to scaffold wrong dependency.